+++ src/main/kotlin/allegrex/agent/ppsspp/bridge/model/PpssppCpuStatus.kt
@@ -3,6 +3,6 @@ package allegrex.agent.ppsspp.bridge.model
 data class PpssppCpuStatus(
   val stepping: Boolean,
   val paused: Boolean,
-  val pc: Int,
-  val ticks: Int,
+  val pc: Long,
+  val ticks: Long,
 )
