+++ src/main/kotlin/allegrex/agent/ppsspp/bridge/model/PpssppHleThread.kt
@@ -1,14 +1,14 @@
 package allegrex.agent.ppsspp.bridge.model
 
 data class PpssppHleThread(
-  val id: Int,
+  val id: Long,
   val name: String,
   val status: Int,
   val statuses: List<String>,
-  val pc: Int,
-  val entry: Int,
-  val initialStackSize: Int,
-  val currentStackSize: Int,
+  val pc: Long,
+  val entry: Long,
+  val initialStackSize: Long,
+  val currentStackSize: Long,
   val priority: Int,
   val waitType: Int,
   val isCurrent: Boolean
