+++ src/main/kotlin/allegrex/agent/ppsspp/bridge/model/PpssppStackFrame.kt
@@ -1,9 +1,9 @@
 package allegrex.agent.ppsspp.bridge.model
 
 data class PpssppStackFrame(
-  val entry: Int,
-  val pc: Int,
-  val sp: Int,
-  val stackSize: Int,
+  val entry: Long,
+  val pc: Long,
+  val sp: Long,
+  val stackSize: Long,
   val code: String
 )
