+++ src/main/kotlin/allegrex/agent/ppsspp/bridge/model/event/PpssppCpuStatusEvent.kt
@@ -5,8 +5,8 @@ import allegrex.agent.ppsspp.bridge.model.PpssppCpuStatus
 data class PpssppCpuStatusEvent(
   val stepping: Boolean,
   val paused: Boolean,
-  val pc: Int,
-  val ticks: Int,
+  val pc: Long,
+  val ticks: Long,
   override val ticket: String?,
 ) : PpssppEvent {
   companion object {
